refactor: #25 공통 뷰 컴포넌트 제작
